+++ tests/test_project.py
@@ -356,16 +356,74 @@ class TestProjectClass(TestProjectBase):
                     project.run()
                 assert os.getcwd() == starting_dir
 
-    def test_cmd_with_job_wrong_order(self):
+    
+    def test_cmd_decorator_with_cmd_argument(self):
         class A(FlowProject):
             pass
 
         with pytest.raises(FlowProjectDefinitionError):
 
+            @A.operation(cmd=True)
+            @cmd
+            def op1(job):
+                pass
+        
+        with pytest.raises(FlowProjectDefinitionError):
+            @cmd
+            @A.operation(cmd=True)
+            def op2(job):
+                pass
+
+    def test_with_job_decorator_with_with_job_argument(self):
+        class A(FlowProject):
+            pass
+
+        with pytest.raises(FlowProjectDefinitionError):
+
+            @A.operation(with_job=True)
+            @with_job
+            def op1(job):
+                pass
+        
+        with pytest.raises(FlowProjectDefinitionError):
+            @with_job
+            @A.operation(with_job=True)
+            def op2(job):
+                pass
+        
+    def test_cmd_with_job_invalid_ordering(self):
+        class A(FlowProject):
+            pass
+
+        with pytest.raises(FlowProjectDefinitionError):
+            @A.operation(cmd=True)
+            @with_job
+            def op1(job):
+                pass
+        
+        with pytest.raises(FlowProjectDefinitionError):
+            @with_job
+            @A.operation(cmd=True)
+            def op2(job):
+                pass
+        
+        with pytest.raises(FlowProjectDefinitionError):
+            @A.operation(with_job=True)
+            @cmd
+            def op3(job):
+                pass
+
+        with pytest.raises(FlowProjectDefinitionError):
+            @cmd
+            @A.operation(with_job=True)
+            def op4(job):
+                pass
+        
+        with pytest.raises(FlowProjectDefinitionError):
             @A.operation
             @cmd
             @with_job
-            def test_cmd(job):
+            def op5(job):
                 pass
 
     def test_with_job_works_with_cmd(self):
